Update 05_efm_pgbouncer.mdx
Missing port number. This worked with edb_pgbouncer and EPAS, but the default port that edb_pgbouncer uses is 5432, while EPAS listens on 5444. The pgbouncer team are considering changing the default, but in the meantime this could result in the customer experience not being the best. And if a customer upgrades pgbouncer to a version that uses 5444 as default, the configuration will still work